imx: Add fsl_sec driver for RNG and identity (#45)
* imx: Add fsl_sec driver for RNG and identity

This fsl_sec driver is ported from the PPA implementation by NXP.

Temporarily included in OP-TEE to provide hardware RNG and OTPMK
derrived identity until a more complete CAAM implementation is ported.

To enable the fsl_sec driver you must set the following configs:
CFG_FSL_SEC=y
CFG_WITH_SOFTWARE_PRNG=n

The RNG calls the CAAM to fill a 256 byte buffer, then iterates through
the buffer to return bytes to callers until it exhausts its pool of
random numbers, at which point it will call the CAAM again. This is to
reduce the performance cost of RNG reads since 16 byte buffers are
requested frequently during RPMB initialization.

The identity derrivation sets the key source to OTPMK and rolls forward
the PRIBLOB to 11 normal mode. This ensures that the identity derived
for RPMB is identical between booting from Cyres SPL and non-Cyres SPL.
If the device isn't in a secure/trusted state it will return a failure
on device identity so that a false RPMB key will not be burned.

Signed-off-by: Henry Beberman <hebeberm@microsoft.com>